requirements.txt
pandas
numpy
spacy
sentence-transformers
bertopic
hdbscan
umap-learn
scikit-learn
kagglehub
openai
python-dotenv
streamlit
pydantic>=2.7
jupyter